Add white top and bottom border around selected row and keep clustercode associated background color.
Add a white top and bottom border around the selected row in the
Transcription Factor and Motif main view en keep de clustercode
associated background color for the selected row (so no blue
selection color) so it is visually easy to see which motifs belong
to the same cluster even when a motif is selected.
The selected row still has white text. To make this white text
visible enough, different colors are used than before.
There are now also 98 different colors instead of 24.