Don't open URL to Transfac Pro matrix multiple times.
Clicking on the Transfac Pro matrix URL resulted in opening
multiple times the same website in the browser.

Each time you switched to another Transfac Pro matrix, clicking
on the link would open the website (n + 2) times (with n = number
of times the URL would be opened for the previous matrix).

Unregistering the mouselistener before registering it, solves
the problem.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>