Make motif and track clusters recognisible by using "M<motifclusternumber>" and "T<trackclusternumber>".
By using "M<motifclusternumber>" for motif clusters and "T<trackclusternumber>"
for track clusters, it is easier to spot which clusters in the TF view are
coming from motifs or tracks.

It also allows to start both motif and track cluster numbering from 1.
For this clusterCode was changed from an integer to a string value,
clusterNumber contains an integer value that was previously in the
clusterCode, so all clusters can be colored in a different color.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>