Add logos for all motifs of motif collection 3 and 6.
Add newly generated logos for all motifs of both motif collection 3 and
motif collection 6. This logos now also have a proper motif name as
title in the PNG image. The logos were generated with convert-matrix of
RSAT tools and the resulting PNG images were reduced as much as possible
with TruePNG and pngzopfli to try to minimize the total size of the PNG
logo files.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>