Catch additional network interupted errors and put the code for creating a connection only in one place.
Put the code for creating a connection based on an URL or a resource bundle key only in
servercommunication.protocols.HTTPProtocol, so that this code is not duplicated anymore.

Also improve the createConnection code so that it will timeout in case the connection
can not be made or in case the connection is interupted, so iRegulon will not hang
anymore in that case.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>