Don't display the "Not a single motif or track is enriched for your input gene signature." error message in case, the connection could not be established or another error occured.
Only display the "Not a single motif or track is enriched for your input gene signature."
error message in case we really didn't get any enriched motifs or tracks (job ran OK).

Display the actual error message to the user instead.
This error messages can be related to problems with the network connection
or server accesibility or with problems of the submitted job itself.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>