Add ClusterColors class so we can get the cluster color based on the cluster number.
Add ClusterColors class so we can get the cluster color based on the cluster number.
Also set the cluster color as an attribute for added edges, so we can use it later
in view.IRegulonVisualStyle for coloring the edges based on the cluster color.
The feature id attribute is not needed anymorem so remove it.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>