Improve iRegulon visual style.
Improve iRegulon visual style:
  - Regulators:
        * shape = octagon
        * color = green
  - Target genes:
        * shape = elipse
        * color = pink
  - Normal genes:
        * shape = elipse
        * color = light blue

The nodes are also a little bit transparent.

All nodes have now also a border in almost the same color as the node itself
(a little bit darker), so when a node is selected (yellow), you can still see
what kind of node it is based on the border.

Edges are show in the same color as the clusters in the tableview (this is not
easily possible with the current API for the arrow edges too).

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>