Fix support for displaying motifs of different motif collections.
Fix support for displaying motifs of different motif collections.

The patch that was supposed to add this support was missing the fact
that motifName.indexOf("__") returns -1 when "__" is not found in
the motifName, so if the logo for a motif of motif collection 3-6
was requested and it did not contain "__", it would try to find
the motif in the motif collection v7 logo directory.

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>